mlapp/pipeline/requirements.txt
opencv-python
ultralytics
onnxruntime
google-cloud-storage
cloud-sql-python-connector
google-cloud-pubsub
psycopg2
numpy
pg8000
python-dotenv